refactor(dpp): split read-only service to fix Turbopack NFT whole-project trace
- 新增 DppSimulatorReadService（getFileTree/getFileDetails/getSimulatorList/deleteSimulatorData，僅碰 data/ + prisma repos）
- list / files 路由改用輕量 read service，不再 import 產生管線模組
- 產生管線 service 一律以 npx tsx 執行 seeder，移除對 node_modules/.bin/tsx symlink 的 path.resolve（NFT 追此 symlink 會拉入 tsx 整包相依，誤判為 whole project）
- 依 CLAUDE.md 單一職責原則拆分，消除 Turbopack build 的 NFT 警告
