+++ src/app/api/v1/digital_product_passport_simulator/list/route.ts
@@ -1,12 +1,12 @@
 import { NextRequest } from "next/server";
 import { jsonOk, jsonFail } from "@/lib/utils/response";
 import { API_ERRORS } from "@/lib/utils/error_dictionary";
-import { DppSimulatorService } from "@/services/dpp_simulator.service";
+import { DppSimulatorReadService } from "@/services/dpp_simulator_read.service";
 
 export const dynamic = "force-dynamic";
 
 export async function GET() {
-  const dppSimulatorService = new DppSimulatorService();
+  const dppSimulatorService = new DppSimulatorReadService();
   const items = await dppSimulatorService.getSimulatorList();
   return jsonOk(items);
 }
@@ -18,7 +18,7 @@ export async function DELETE(req: NextRequest) {
       return jsonFail(API_ERRORS.VL_MISSING_PARAMS);
     }
 
-    const dppSimulatorService = new DppSimulatorService();
+    const dppSimulatorService = new DppSimulatorReadService();
     await dppSimulatorService.deleteSimulatorData(stockId, year);
 
     return jsonOk(null);
